Add files via upload
This heatmap was created for the purpose of inclusion in an MSc thesis 'Applied voltage frequency concerning ion power deposition ionization rates and on energy distribution functions.
The 3D heatmap displays surface impact energies with integrated angles of inclination as colour intensity. The surface in this instance is an alumina dielectric encompassing a powered electrode, primarily for an asymmetric capacitively coupled parallel plate radio frequency electrothermal micro-thruster. But a nice example of a heatmap disregarding the specifics.